# apps/api-gateway/src/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EduFlow API Documentation",
    "description": "Comprehensive API documentation for all EduFlow services",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/kyc/api/kyc/documents": {
      "post": {
        "tags": [
          "kyc",
          "documents"
        ],
        "description": "Submit a new KYC document for verification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "documentUrls",
                  "metadata"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "IDENTITY",
                      "SCHOOL_LICENSE",
                      "EMPLOYMENT_PROOF",
                      "QUALIFICATION"
                    ]
                  },
                  "documentUrls": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uri"
                    }
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Document submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Document submitted successfully",
                  "type": "object",
                  "required": [
                    "id",
                    "userId",
                    "type",
                    "status",
                    "documentUrls",
                    "metadata",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "userId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "IDENTITY",
                        "SCHOOL_LICENSE",
                        "EMPLOYMENT_PROOF",
                        "QUALIFICATION"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "VERIFIED",
                        "REJECTED",
                        "EXPIRED"
                      ]
                    },
                    "documentUrls": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "uri"
                      }
                    },
                    "metadata": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "verifiedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Invalid request body",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Unauthorized",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "kyc",
          "documents"
        ],
        "description": "Get all documents for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user documents",
            "content": {
              "application/json": {
                "schema": {
                  "description": "List of user documents",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "IDENTITY",
                          "SCHOOL_LICENSE",
                          "EMPLOYMENT_PROOF",
                          "QUALIFICATION"
                        ]
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "PENDING",
                          "VERIFIED",
                          "REJECTED",
                          "EXPIRED"
                        ]
                      },
                      "documentUrls": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "format": "uri"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kyc/api/kyc/documents/{documentId}/verify": {
      "patch": {
        "tags": [
          "kyc",
          "verification"
        ],
        "description": "Verify a KYC document",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "PENDING",
                      "VERIFIED",
                      "REJECTED",
                      "EXPIRED"
                    ]
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "documentId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Document verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Document verified successfully",
                  "type": "object",
                  "required": [
                    "id",
                    "status",
                    "verifiedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "VERIFIED",
                        "REJECTED",
                        "EXPIRED"
                      ]
                    },
                    "verifiedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kyc/api/kyc/history/{entityId}": {
      "get": {
        "tags": [
          "verification"
        ],
        "description": "Get verification history for an entity",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "entityId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Verification history",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Verification history",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "status",
                      "verifiedBy",
                      "createdAt"
                    ],
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "PENDING",
                          "VERIFIED",
                          "REJECTED",
                          "EXPIRED"
                        ]
                      },
                      "verifiedBy": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "notes": {
                        "type": "string",
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}